requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
pytest==7.4.3
python-dotenv==1.0.0
python-json-logger==2.0.7
redis==5.0.1
requests==2.31.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
